product_research_engine/requirements.txt
pandas
numpy
beautifulsoup4
requests
pyyaml
rapidfuzz
lxml
playwright
